Refactor category selector
